requirements.txt
# Core dependencies
python-binance
ccxt==4.0.112
pandas==2.2.0
numpy==1.26.3
ta==0.11.0
TA-Lib

# AI/ML dependencies
openai==1.3.0
anthropic==0.18.1
transformers==4.37.2
torch==2.2.0

# Sentiment analysis
tweepy==4.14.0
praw==7.7.1
newspaper4k
textblob==0.18.0
vaderSentiment==3.3.2

# Data and visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.2

# Telegram bot
python-telegram-bot==20.6

# Web scraping and APIs
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.17.2
aiohttp==3.9.3

# Database
sqlalchemy==2.0.25
redis==5.0.1

# Utils
python-dotenv==1.0.1
pydantic==2.5.3
loguru==0.7.2
schedule==1.2.0
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# API Server
flask==3.0.0